gestion upload ok

+++ backend/controllers/user/filesController.js
@@ -1,3 +1,4 @@
+import { createCipheriv } from "crypto";
 import formidable from "formidable";
 import fs from "fs";
 
@@ -5,29 +6,41 @@ export const uploadFile = (req, res) => {
   const form = formidable();
   form.parse(req, function (error, fields, files) {
     try {
-      fs.unlink(fields.path, (error) => {
-        if (error) {
-          return error;
-        }
-        console.log("Le fichier a été supprimé avec succès");
-      });
-      const projectId = fields.projectId;
-      const extension = files.file.originalFilename.split(".").pop();
-      let oldpath = files.file.filepath;
-      let newpath =
-        `public/images/${projectId}_` +
-        files.file.newFilename +
-        "." +
-        extension;
-      fs.copyFile(oldpath, newpath, (error) => {
-        if (error) {
-          return error;
-        }
-        res.status(200).json({ newpath });
-      });
+      if (!files.file) {
+        res.status(400).json({
+          message: "fichier manquant",
+        });
+      } else {
+        const projectId = fields.projectId;
+        const extension = files.file.originalFilename.split(".").pop();
+        let oldpath = files.file.filepath;
+        let newpath =
+          `public/images/${projectId}_` +
+          files.file.newFilename +
+          "." +
+          extension;
+        fs.copyFile(oldpath, newpath, (error) => {
+          if (error) {
+            res.send(error);
+          } else {
+            if (fields.path) {
+              fs.unlink(fields.path, (error) => {
+                if (error) {
+                  return res.send(error);
+                } else {
+                  console.log("fichier supprimé");
+                  res.status(200).json({ newpath });
+                }
+              });
+            } else {
+              res.status(200).json({ newpath });
+            }
+          }
+        });
+      }
     } catch (error) {
       res.status(400).json({
-        message: "Erreur de l'upload",
+        message: "upload impossible",
         erreur: error,
       });
     }
